Make util test function part of test files
The function defined in the testing.go file is only part of the tests
and it is importing "testing".
Adding _test suffix makes it part of the test suite only so the function
is not part of the production code, which is also not importing
"testing" package in production binary as well.